ajuste

+++ bingo-api/src/DTOs/Response/Scratch/Jsonb/ScratchSymbolDto.cs
@@ -5,7 +5,7 @@ namespace bingo_api.src.DTOs.Response.Scratch.Jsonb;
 public record ScratchSymbolDto
 {
     public string Symbol { get; set; }
-    public decimal PrizeValue { get; set; }
+    public double PrizeValue { get; set; }
 
     internal static ScratchSymbolDto ConvertToDto(ScratchSymbol symbol)
     {
